fix: make beaten_distances floats

+++ src/models/formdata_run.py
@@ -1,4 +1,3 @@
-from decimal import Decimal
 from typing import Optional
 
 from pydantic import BaseModel, Field
@@ -15,7 +14,7 @@ class FormdataRun(BaseModel):
     allowance: Optional[int] = None
     jockey: str
     position: str
-    beaten_distance: Optional[Decimal] = None
+    beaten_distance: Optional[float] = None
     time_rating: Optional[int] = None
     distance: float
     going: str
