+++ src/models/formdata_runner.py
@@ -1,4 +1,3 @@
-from decimal import Decimal
 from typing import Optional
 
 from pydantic import BaseModel, Field
@@ -13,6 +12,6 @@ class FormdataRunner(BaseModel):
     allowance: Optional[int] = None
     jockey: str
     position: str
-    beaten_distance: Optional[Decimal] = None
+    beaten_distance: Optional[float] = None
     time_rating: Optional[int] = None
     form_rating: Optional[int] = None
